REFACTOR :: 구조 변경
