--- apps/user/src/app/form/전형선택/hooks/useFormTypeProvider.ts
@@ -1,12 +1,0 @@
-import { atom, useRecoilState } from 'recoil';
-
-const formTypeAtomState = atom({
-    key: 'form-type',
-    default: '',
-});
-
-export const useFormTypeProvider = () => {
-    const [formType, setFormType] = useRecoilState(formTypeAtomState);
-
-    return { formType, setFormType };
-};
