+++ apps/user/src/components/form/ProfileUpload/ProfileUpload.tsx
@@ -9,7 +9,7 @@ import {
     useRef,
     DragEvent,
 } from 'react';
-import { UserInfo } from '@/app/form/지원자정보/hooks/useUserInfoProvider';
+import { UserInfo } from '@/app/form/지원자정보/지원자정보.provider';
 import { useUploadProfileImageMutation } from '@/services/form/지원자정보/mutations';
 import styled from 'styled-components';
